Programa corrigido

+++ include/relays.h
@@ -18,6 +18,6 @@ void valveON(const int valve);
 void valveOFF(const int valve);
 
 void tempControl(const int hotLight, const int fan, const float average, const float max, const float min);
-void humControl(const int valve, const int average, const int max);
+void humControl(const int valve, const float average, const float max);
 
 #endif
